pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-tour-guide"
version = "1.0.0"
description = "Multi-agent system that enriches walking tours with multimedia content"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "googlemaps>=4.10.0",
    "google-api-python-client>=2.108.0",
    "spotipy>=2.23.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
tour-guide = "tour_guide:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pylint.main]
fail-under = 8.0

[tool.pylint."messages control"]
disable = [
    "logging-fstring-interpolation",  # f-strings in logging are fine for this project
    "too-many-instance-attributes",
    "too-many-arguments",
    "too-many-positional-arguments",
    "broad-exception-caught",         # agents intentionally catch broad exceptions
    "duplicate-code",                 # test fixtures share setup code by design
    "redefined-outer-name",           # pytest fixtures redefine outer names
    "import-outside-toplevel",        # conditional imports in tests
    "wrong-import-order",             # not critical
    "protected-access",               # tests access private methods for unit testing
    "use-implicit-booleaness-not-comparison",
]

[tool.pylint.format]
max-line-length = 120
